Final official commit
Well, I optimized it so it corresponded to the assignment. Plus, it's a
lot more readable (removed test prints and polished up commentary). Was
a very fun project, and at this state is still open to a lot of
features, which I may or may not add in the near future :D